tdd progress on lineTracer

+++ Source/Voxel/LineTracer.h
@@ -5,6 +5,6 @@
 
 class LineTracer {
 public:
-  static std::vector<std::vector<glm::dvec3>> traceLines(const bool* data, const glm::ivec3& resolution); 
+  static std::vector<std::vector<glm::ivec3>> traceLines(const bool* data, const glm::ivec3& resolution); 
   static std::vector<glm::ivec3> gatherTrues(const bool* data, const glm::ivec3& resolution);
 };
